sources/requirements.txt
httplib2
Pillow
PyGObject
requests
youtube_search_python
yt_dlp
